Add files via upload
The development and adjustment of the home page has been completed. The pictures and logo are waiting to be changed. The navigation bar has been completed. I am considering whether to use anchor navigation or website navigation.
Referring to a wave particle display on the Internet as the background of the home page,