Add files via upload
The view logic is updated. Now users can provide data sources for other people's evidence. Note: the rules are set. If the case of this evidence belongs to the current user, the source can be added directly. That is, the user is adding a source for the evidence provided by others in the case that he owns, and the user has autonomy
If the case of this evidence does not belong to the current user,then the user needs to verify evidence owner's key(draft)