Add files via upload
The last code update, including 1. You can add attachments to the source 2. The pattern page content update 3. Intelligent automatic analysis and the realization of the case, evidence, and source respectively 4. The last update of the model, the last update of the database 5. The overall analysis report of the case
All basic functions have been implemented